pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "chembl-classify"
version = "0.1.0"
dependencies = [
    "pandas>=2.0",
    "pyyaml>=6.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "W", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
